basic functionality added
Created a very basic game scene where a player object and a set number of coin objects are spawned by a GameManager script attached to an empty object.

Added a MainManager script to handle data persistence between scenes (as in the previous projects), attached to an empty object in the same scene.

Inside this script, an instance of MainManager is created (tried basic encapsulation there with get/set).
This script also stores an int var named score.

On collision (compareTag "Coin"), a PlayerController script attached to the player object counts up MainManager.Instance.score += 10, and a local score text UI is updated with that value via the GameManager script.

Added a basic shop scene, which the player can enter from the game scene by pressing the E key.

In this new scene, a ShopManager script is attached to an empty object.
This script spawns one of three available items and sets an associated UI text active.

Each item (banana, apple, supply crate) inherits from an Items class
(tried basic overriding here - two of them override a virtual Move() method of that parent class).

Added functionality to browse the items with the horizontal Arrow keys (prev/next), and "buy" an item by pressing the B key
(the score var is stored and passed over from the game scene with the MainManager).

Each item has a different int value price assigned to it in the ShopManager script.

When an item is bought, MainManager.Instance.score is reduced by price in the ShopManager script and a local score text UI is updated by it to show the resulting score.

Pressing the E key allows one to go back to/loads the game scene.

The current (reduced) score is once again stored and passed back over to the game scene by the MainManager.

That´s pretty much  it.

Hope to do some branching next...