title screen added
Created a startmenu scene with a title screen canvas containing a title text and a button.

Removed the MainManager empty containing the MM script from the game scene and created it new in the startmenu scene (now scene 0).

Added a public StartGame method to the MainManager script. This is called when the new button is clicked and loads the game scene (scene 1).

Fixed some minor RBD issues.