encapsulated the score variable
Encapsulated the score var using a private field m_score.

Added an if-else statement to the set, preventing a negative score value.

Added a Coin script to the coin pick up, which sets up a random Y rotation in Start and rotates the coin around the Y axis in Update.

Started adding "Pillars" comments/notes to the code ("//ENCAPSULATION", "//INHERITANCE" etc.).

Created a basic shopping cart geo in Blender and imported it into the project, for a possible branch featuring an energy bar and hazards/"enemies".