animator added/title screen minor changes
Removed the placeholder object formerly used for the player character.

Imported an animated model from Blender
(been doing tests with automated keyframing via Python and creating animations with the sine and cosine functions that way - this is just one of the first sine/cosine walk cycles and sine/cosine idle anims).

In Unity, created a prefab from the imported FBX, added an animator controller, and set up the idle and walk animations in the animator tab/window.

Accessing the animator controller in the PlayerController script with playerAnim = GetComponent and controlling the transitions (idle - walk and back) via a bool parameter idle_b with playerAnim.SetBool().

Changed the movement controls in the PlayerController:

up and down Arrow keys now move the character forward/backwards whilst left/right Arrow keys rotate the character around the Y axis.

Created two renders of the shopping cart geo in different colors, imported them into Unity and added  them to the title screen to make it a bit more appealing.

That´s pretty much it.