energy bar plus functionality added
Added two UI images to the player UI in the game scene, one for the energy bar background, one for its content.

Added a new var energy to the MainManager with the same encapsulation as score (no negative values).

Accessing the energy bar content image in the GameManager and adjusting the transforms/width based on the value of MainManager.Instance.energy.

Since there are no hazards/"enemies" yet, tested it by subtracting MainManager.Instance.energy -= 10 in the OnCollisionEnter of the PlayerController when a coin pick up is collected .

Works fine.