hazard/enemy functionality added
Created an Enemy script and attached it to the shopping cart prefab.

Inside this script, added a MoveEnemy method which transform.Translates along Vector3.forward by Time.deltaTime and a private speed var.

This method is then called in Update, together with a likewise new UpdateRotation method, which in turn checks the X position and transform.Rotates the "enemy"  by 180 deg, if posX is greater/less than a set value, making the object "patrol" left/right.

Set the shopping cart prefab´s box collider to act as a trigger and added an OnTriggerEnter to the Enemy script.

If other.CompareTag is "Player", energy is subtracted with MainManager.Instance.energy -= 10.

Works fine.