items replenish energy functionality added
Recreated the energy bar from the game scene in the shop scene and set it up in ShopManager to reflect MainManager.Instance.energy, inside a new UpdateEnergyBar method, which is then called in Update of ShopManager.

Added a new method SetEnergyGain to ShopManager, which sets a private energy var to a given value on each item.

On purchasing an item by pressing the B key, the BuyItem method is called as before.
Additionally to calling SetPrice and subtracting MainManager.Instance.score -= price, it now also calls SetEnergyGain and adds MainManager.Instance.energy += energy.

That´s pretty much it.