basic inventory canvas plus toggle added
Created a new canvas named Inventory in the game scene.

Added images for the items with nested texts to it.

Created a new method ToggleInventory in GameManager, which sets a new private bool var isInvetoryActive to true/false and likewise unhides/hides the inventory with inventory.gameObject.SetActive.

This method is then called in the Update of GamManager when the I (i) key is pressed.

That´s pretty much it.