very basic inventory implemented
Finished implementing the "inventory".

Items purchased in the shop scene are now counted up there by the ShopManager and their count is written into/processed by the MainManager, just like score/energy.

The items´ respective counts are displayed in a new inventory canvas (copied the finished one from the game scene).

Back in the game scene, purchased items (banana, apple, crate) can now be used with the alphanumeric keys 1 to 3 to replenish energy (update the energy var/bar) according to the energy gain values of the different items. Using an item reduces its count by 1.

The enegy gain values of the items are now encapsulated in the MainManager with get private set, and each is assigned a different value in Start there.

That´s pretty much it.