incrementing enemy count on scene reload added
Since the GameManager script handles all the spawning in the game scene, added a private static int var enemyCount there, setting it to 1 on declaration.

This var is incremented in the SpawnEnemies method called in Start and, since it´s static, is persistent between reloads of the scene.

Now each time the game scene is re-entered from the shop scene, an additional hazard/"enemy" is spawned, until the enemyCount reaches 5 in total (else it´s just too crowded).

That´s pretty much it.